// src/utils/types/cartItem.ts
  export interface CartItem {
    productId: string;
    quantity: number;
    title: string;
    imageURL: string;
    price: number;
  }

  // const cartItem = {
  //   productId: product.id,
  //   quantity: 1,
  //   title: product.title, // Add title property
  //   imageURL: product.imageURL, // Add imageURL property
  //   price: product.price, // Add price property
  // };